Fixed syntac issues and verified code compilation

+++ firmware/STM32 Drivers/MCP_3221/Core/Inc/viper_mcp3221.h
@@ -13,10 +13,10 @@
 #include "mcp_3221.h"
 
 // Define VIPER-specific constants
-#define VIPER_SCALING_FACTOR        double(17.6)      // Voltage divider scaling factor
-#define VIPER_LOW_SENSE_RES         double(0.002)     // Sense resistor value in ohms for low power card
-#define VIPER_HIGH_SENSE_RES        double(0.001)     // Sense resistor value in ohms for high power card
-#define VIPER_VREF_MV               double(3300.0)    // Reference voltage in millivolts
+#define VIPER_SCALING_FACTOR        (17.6)      // Voltage divider scaling factor
+#define VIPER_LOW_SENSE_RES         (0.002)     // Sense resistor value in ohms for low power card
+#define VIPER_HIGH_SENSE_RES        (0.001)     // Sense resistor value in ohms for high power card
+#define VIPER_VREF_MV               (3300.0)    // Reference voltage in millivolts
 #define VIPER_ADC_ADDRESS           0x4D    		 // I2C address for the MCP3221 device
 
 // Function declarations
